Add encode_jwt and decode_jwt methods to AttestationResult class
