+++ src/claims.py
@@ -1,9 +1,13 @@
 import json
 from dataclasses import dataclass, field
+from datetime import datetime, timedelta
 from typing import Any, Dict
+
+from jose import jwt  # type: ignore # pylint: disable=import-error
 
 from src.base import BaseJCSerializable
 from src.errors import EARValidationError
+from src.jwt_config import DEFAULT_ALGORITHM, DEFAULT_EXPIRATION_MINUTES
 from src.trust_tier import to_trust_tier
 from src.trust_vector import TrustVector
 from src.verifier_id import VerifierID
@@ -116,3 +120,27 @@ class AttestationResult(BaseJCSerializable):
 
             trust_vector = details["trust_vector"]
             trust_vector.validate()
+
+    def encode_jwt(
+        self,
+        secret_key: str,
+        algorithm: str = DEFAULT_ALGORITHM,
+        expiration_minutes: int = DEFAULT_EXPIRATION_MINUTES,
+    ) -> str:
+        # Signs an AttestationResult object and returns a JWT
+        payload = self.to_dict()
+        payload["exp"] = int(
+            datetime.timestamp(datetime.now() + timedelta(minutes=expiration_minutes))
+        )
+        return jwt.encode(payload, secret_key, algorithm=algorithm)
+
+    @classmethod
+    def decode_jwt(
+        cls, token: str, secret_key: str, algorithm: str = DEFAULT_ALGORITHM
+    ):
+        # Verifies a JWT and returns the decoded AttestationResult object.
+        try:
+            payload = jwt.decode(token, secret_key, algorithms=[algorithm])
+            return cls.from_dict(payload)
+        except Exception as exc:
+            raise ValueError(f"JWT decoding failed: {exc}") from exc
