+++ tox.ini
@@ -10,6 +10,7 @@ deps =
     pylint==2.17.5
     pyright==1.1.325
     pytest==7.4.2
+    python-jose==3.4.0
 commands =
     isort . --profile=black
     black . --check --diff
@@ -21,4 +22,5 @@ commands =
 [testenv:test]
 deps =
     pytest==7.4.2
+    python-jose==3.4.0
 commands = pytest
